Add support for connection over websocket for relays behind proxy.
This adds a major milestone for this application. We now support
connecting over the websocket connection to talk to a relay behind
a reverse proxy  (e.g. Nginx in my case) and doesn't require a ssh
connection with reverse proxy.

We also support connecting the old way, although, I haven't yet
added a way to choose between them. It can be flipped by changing
the input to WeechatConnFactory() method in main.go.

We also support notifications. It has been tested on macOS so far.
There should be a way to configure somewhere the words that trigger
a notification. That requires support for a proper CLI and configuration
mechanism which we don't currently have.